\name{gstart}
\alias{gstart}

\title{Generate approximate gamma distribution parameters.}

\description{This function generates parameters suitable for starting a likelihood optimization.
}
\usage{
gstart(x)
}

\arguments{
  \item{x}{A dataframe such as generated by mleframe with column names 'left', 'right' and optionally 'qty'.
    Exact failure data (occurrences) have same time entered in both 'left' and 'right' columns.
	Suspension data has last known life[time] entered in 'left' column and -1 entered in 'right' column.
	The left(early) interval bound for left-censored data must be entered as zero. (NA is not accepted).
  }
}

\value{
  Returns a vector of alpha (shape) and beta (rate) parameters for the gamma distribution.
}

\details{
  This function performs a correlation of the gamma shape parameter based on the weibull shape parameter fitted to the data.  The beta (rate) parameter is then determined utilizing the mean of failure data.
  The correlation is significantly affected by the occurence of right-censored, suspension, data. This correlation currently considers these suspensions to be evenly occurring within the data. Strong alteration is known to exist in the 
 presence of late suspensions, which has not yet been modeled.  
}

\references{
  William Q. Meeker and Luis A. Escobar, (1998) "Statistical Methods for Reliability Data", Wiley-Interscience, New York

  Robert B. Abernethy, (2008) "The New Weibull Handbook, Fifth Edition" 

  Marie Laure Delignette-Muller, Christophe Dutang (2015). "fitdistrplus: An R Package for Fitting Distributions".
  Journal of Statistical Software, 64(4), 1-34. URL http://www.jstatsoft.org/v64/i04/.
}

\examples{
failures<-c(90,96,30,49,82)
suspensions<-c(100,45,10)
gstart(mleframe(failures,suspensions))
}
