\name{mle_gamma}
\alias{mle_gamma}

\title{Maximum likelihood regression on the gamma distribution}

\description{
 Maximum likelihood regression on the gamma distribution.
}

\usage{
mle_gamma(x, par,  control=NULL) 
}

\arguments{
  \item{x}{A dataframe such as generated by mleframe with column names 'left', 'right' and optionally 'qty'.
    Exact failure data (occurrences) have same time entered in both 'left' and 'right' columns.
	Suspension data has last known life[time] entered in 'left' column and -1 entered in 'right' column.
	The left(early) interval bound for left-censored data must be entered as zero. (NA is not accepted).
  }
  \item{par}{A vector holding the approximation of the gamma shape and rate parameters. This approximation is expected to be derived from function gstart.
  }
  \item{control}{An optional list suitable for passing to the optim function.}  
}

\value{
Returns a named vector with parameters of the distribution fit and goodness of fit in terms of log likelihood.
}

\details{
The negative log-likelihood is minimized by the Nelder-Meade, simplex, algorithm.  This algorithm requires a reasonable starting point
for the estimate.  The mle_gamma function makes a call to optim in the R base rather than custom code for the simplex algorithm as used by mlefit in WeibullR.
 The control argument has been provided to give debug developers a means of altering and examining the progress of the optimization.
}

\references{
  William Q. Meeker and Luis A. Escobar, (1998) "Statistical Methods for Reliability Data", Wiley-Interscience, New York
  
  Marie Laure Delignette-Muller, Christophe Dutang (2015). "fitdistrplus: An R Package for Fitting Distributions".
  Journal of Statistical Software, 64(4), 1-34. URL http://www.jstatsoft.org/v64/i04/.
}

\examples{
failures<-c(90,96,30,49,82)
suspensions<-c(100,45,10)
mle_data<-mleframe(failures, suspensions)
gamma_fit<-mle_gamma(mle_data, gstart(mle_data))
}

\keyword{ regression reliability }